# apps/api/src/provider-source-routes.ts
import type{FastifyInstance,FastifyRequest}from'fastify';import type{LocalAuthService}from'@scoutops/auth';import type{AuthorizationService}from'@scoutops/authorization';import type{ProviderSourceService}from'./provider-source-service.js';import{sessionToken}from'./auth-routes.js';import{ApiError,requireIdempotencyKey}from'./api-foundation.js';
export interface ProviderSourceRouteOptions{service:ProviderSourceService;authorization:AuthorizationService;auth:LocalAuthService;secureCookie:boolean;webOrigin:string;}const ids=(r:FastifyRequest)=>({requestId:r.headers['x-request-id']!.toString(),traceId:r.headers['x-trace-id']!.toString()}),envelope=(data:unknown,r:FastifyRequest)=>({data,request_id:ids(r).requestId,trace_id:ids(r).traceId});
export function registerProviderSourceRoutes(app:FastifyInstance,o:ProviderSourceRouteOptions){const actor=async(r:FastifyRequest,capability:'provider:configure'|'collection:replay'|'trend:read')=>{const a=await o.auth.authenticate(sessionToken(r,o.secureCookie));await o.authorization.authorize({actorId:a.user.id,capability,surface:'api',...ids(r)});return a.user.id;},write=async(r:FastifyRequest,capability:'provider:configure'|'collection:replay'|'trend:read')=>{if(r.headers.origin!==o.webOrigin)throw new ApiError(403,'origin_forbidden','请求来源不允许。','从 ai选品 页面重试。');return actor(r,capability);};app.get('/api/v1/platform/provider-sources',async(r,reply)=>{await actor(r,'provider:configure');reply.header('cache-control','private, no-store');return envelope(await o.service.list(),r);});app.post('/api/v1/platform/provider-sources/:code/provision',async(r,reply)=>{const actorId=await write(r,'provider:configure'),result=await o.service.provision((r.params as{code:string}).code,{actorId,idempotencyKey:requireIdempotencyKey(r),...ids(r)});reply.code(201);return envelope(result,r);});app.post('/api/v1/platform/provider-sources/:providerId/replays',async(r,reply)=>{const actorId=await write(r,'collection:replay'),result=await o.service.replay((r.params as{providerId:string}).providerId,r.body as any,{actorId,idempotencyKey:requireIdempotencyKey(r),...ids(r)});reply.code(202);return envelope(result,r);});app.post('/api/v1/provider-sources/refresh',async(r,reply)=>{const actorId=await write(r,'trend:read'),result=await o.service.refresh(r.body as any,{actorId,idempotencyKey:requireIdempotencyKey(r),...ids(r)});reply.code(202);return envelope(result,r);});}