# apps/api/src/provider-source-service.ts
import { createHash, randomUUID } from 'node:crypto';
import { BUILTIN_PROVIDER_SOURCES, parseProductSupplyCsv, type BuiltinSourceDefinition } from '@scoutops/provider-sources';

export class ProviderSourceServiceError extends Error{constructor(readonly code:string,readonly statusCode:number,readonly actionHint:string){super(code);this.name='ProviderSourceServiceError';}}
export interface ProvisionedSource{id:string;code:string;status:'draft'|'disabled'|'enabled';version:number;updated_at:string;}
export interface ProviderSourceCatalogItem extends BuiltinSourceDefinition{provisioned:ProvisionedSource|null;}
export interface ProviderSourceReplay{id:string;task_id:string;provider_id:string;source_code:string;status:string;item_count:number;error_code:string|null;request_id:string;trace_id:string;created_at:string;updated_at:string;}
export interface ProviderSourceRepository{
  listProvisioned(codes:string[]):Promise<ProvisionedSource[]>;
  syncCatalog(input:{definitions:readonly BuiltinSourceDefinition[];now:Date}):Promise<{inserted:number;updated:number;automatic_enabled:number;status:'synced'|'waiting_for_platform_admin'}>;
  provision(input:{definition:BuiltinSourceDefinition;providerId:string;actorId:string;route:string;idempotencyKey:string;requestId:string;traceId:string;now:Date}):Promise<ProvisionedSource>;
  replay(input:{providerId:string;sourceCode:string;organizationId:string;workspaceId:string;target:Record<string,unknown>;inputSha256:string;runId:string;taskId:string;subqueryId:string;actorId:string;route:string;idempotencyKey:string;requestId:string;traceId:string;now:Date}):Promise<ProviderSourceReplay>;
  refresh(input:{organizationId:string;workspaceId:string;actorId:string;route:string;idempotencyKey:string;requestId:string;traceId:string;now:Date}):Promise<{task_id:string;source_count:number;status:'scheduled'}>;
}
const uuid=/^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
const source=(code:string)=>{const result=BUILTIN_PROVIDER_SOURCES.find(item=>item.code===code);if(!result)throw new ProviderSourceServiceError('provider_source_not_found',404,'刷新首批来源目录。');return result;};
export class ProviderSourceService{
  constructor(private readonly repository:ProviderSourceRepository,private readonly now:()=>Date=()=>new Date()){}
  async list():Promise<ProviderSourceCatalogItem[]>{const provisioned=await this.repository.listProvisioned(BUILTIN_PROVIDER_SOURCES.map(item=>item.code));return BUILTIN_PROVIDER_SOURCES.map(item=>({...item,provisioned:provisioned.find(value=>value.code===item.code)??null}));}
  ensureCatalog(){return this.repository.syncCatalog({definitions:BUILTIN_PROVIDER_SOURCES,now:this.now()});}
  provision(code:string,context:{actorId:string;idempotencyKey:string;requestId:string;traceId:string}){const definition=source(code);return this.repository.provision({definition,providerId:randomUUID(),route:`/platform/provider-sources/${code}/provision`,...context,now:this.now()});}
  replay(providerId:string,value:{organization_id?:unknown;workspace_id?:unknown;query?:unknown;csv_text?:unknown},context:{actorId:string;idempotencyKey:string;requestId:string;traceId:string}){if(!uuid.test(providerId))throw new ProviderSourceServiceError('provider_id_invalid',400,'刷新首批来源目录。');if(!uuid.test(String(value?.organization_id??''))||!uuid.test(String(value?.workspace_id??'')))throw new ProviderSourceServiceError('provider_source_scope_invalid',400,'选择有效的组织与工作区。');let target:Record<string,unknown>;if(value.query!==undefined&&value.csv_text!==undefined)throw new ProviderSourceServiceError('provider_source_target_invalid',400,'query 与 csv_text 只能提交一种。');if(value.query!==undefined){if(typeof value.query!=='string'||!value.query.trim()||value.query.length>200)throw new ProviderSourceServiceError('provider_source_query_invalid',400,'关键词需要 1–200 字符。');target={query:value.query.trim()};}else if(value.csv_text!==undefined){if(typeof value.csv_text!=='string')throw new ProviderSourceServiceError('provider_source_csv_invalid',400,'CSV 内容格式无效。');try{parseProductSupplyCsv(value.csv_text,20);}catch(error){throw new ProviderSourceServiceError('provider_source_csv_invalid',400,error instanceof Error?error.message:'按固定表头修正 CSV。');}target={csv_text:value.csv_text};}else throw new ProviderSourceServiceError('provider_source_target_invalid',400,'提交关键词或 CSV 内容。');const serialized=JSON.stringify(target);return this.repository.replay({providerId,sourceCode:'',organizationId:String(value.organization_id),workspaceId:String(value.workspace_id),target,inputSha256:createHash('sha256').update(serialized).digest('hex'),runId:randomUUID(),taskId:randomUUID(),subqueryId:randomUUID(),route:`/platform/provider-sources/${providerId}/replays`,...context,now:this.now()});}
  refresh(value:{organization_id?:unknown;workspace_id?:unknown},context:{actorId:string;idempotencyKey:string;requestId:string;traceId:string}){const organizationId=String(value?.organization_id??''),workspaceId=String(value?.workspace_id??'');if(!uuid.test(organizationId)||!uuid.test(workspaceId))throw new ProviderSourceServiceError('provider_source_scope_invalid',400,'重新选择组织和工作区后再刷新热点。');return this.repository.refresh({organizationId,workspaceId,route:'/provider-sources/refresh',...context,now:this.now()});}
}